Added TemperaturePoints + full amazing relatedStuff, like repository, service and implementation, is's like an Entity now!

+++ src/main/java/com/greenbox/test/demo/controller/form/convert/GreenBoxRegistrationForm.java
@@ -11,6 +11,6 @@ public class GreenBoxRegistrationForm {
     @NotNull
     private String name;
     private Long userId;
-    private Integer growProgramId;
+    private Long growProgramId;
 
 }
